refactor features-test command and stub out features-package
