+++ .gitignore
@@ -5,3 +5,4 @@ logs
 *.tgz
 tmp
 .DS_Store
+.env
