+++ .vscode/launch.json
@@ -10,10 +10,11 @@
             "name": "Launch CLI - up",
             "program": "${workspaceFolder}/src/spec-node/devContainersSpecCLI.ts",
             "cwd": "${workspaceFolder}",
+            "envFile": "${workspaceFolder}/.env",
             "args": [
                 "up",
                 "--workspace-folder",
-                "../devcontainers-features",
+                "/home/node/example-project",
                 "--log-level",
                 "debug",
             ],
