+++ src/spec-node/devContainersSpecCLI.ts
@@ -21,11 +21,11 @@ import { getDockerComposeFilePaths } from '../spec-configuration/configuration';
 import { workspaceFromPath } from '../spec-utils/workspaces';
 import { readDevContainerConfigFile } from './configContainer';
 import { getDefaultDevContainerConfigPath, getDevContainerConfigPathIn, uriToFsPath } from '../spec-configuration/configurationCommonUtils';
-import { CLIHost, getCLIHost } from '../spec-common/cliHost';
+import { getCLIHost } from '../spec-common/cliHost';
 import { loadNativeModule } from '../spec-common/commonUtils';
 import { generateFeaturesConfig, getContainerFeaturesFolder } from '../spec-configuration/containerFeaturesConfiguration';
-import { doFeaturesTestCommand } from './featuresCLI/testContainerFeatures';
-import { PackageConfiguration } from '../spec-utils/product';
+import { featuresTestOptions, featuresTestHandler } from './featuresCLI/test';
+import { featuresPackageHandler, featuresPackageOptions } from './featuresCLI/package';
 
 const defaultDefaultUserEnvProbe: UserEnvProbe = 'loginInteractiveShell';
 
@@ -49,14 +49,16 @@ const defaultDefaultUserEnvProbe: UserEnvProbe = 'loginInteractiveShell';
 	y.command('build [path]', 'Build a dev container image', buildOptions, buildHandler);
 	y.command('run-user-commands', 'Run user commands', runUserCommandsOptions, runUserCommandsHandler);
 	y.command('read-configuration', 'Read configuration', readConfigurationOptions, readConfigurationHandler);
-	y.command('features', 'Features commands', (y: Argv) =>
-		y.command('test', 'Test features', featuresTestOptions, featuresTestHandler));
+	y.command('features', 'Features commands', (y: Argv) => {
+		y.command('test', 'Test features', featuresTestOptions, featuresTestHandler);
+		y.command('package', 'Package features', featuresPackageOptions, featuresPackageHandler);
+	});
 	y.command(restArgs ? ['exec', '*'] : ['exec <cmd> [args..]'], 'Execute a command on a running dev container', execOptions, execHandler);
 	y.parse(restArgs ? argv.slice(1) : argv);
 
 })().catch(console.error);
 
-type UnpackArgv<T> = T extends Argv<infer U> ? U : T;
+export type UnpackArgv<T> = T extends Argv<infer U> ? U : T;
 
 const mountRegex = /^type=(bind|volume),source=([^,]+),target=([^,]+)(?:,external=(true|false))?$/;
 
@@ -831,85 +833,6 @@ export async function doExec({
 		};
 	}
 }
-
-// -- 'features test' command
-function featuresTestOptions(y: Argv) {
-	return y
-		.options({
-			'base-image': { type: 'string', alias: 'i', default: 'ubuntu:focal', description: 'Base Image' },
-			'features': { type: 'array', alias: 'f', describe: 'Feature(s) to test as space-separated parameters. Omit to auto-detect all features in collection directory.  Cannot be combined with \'-s\'.', },
-			'scenarios': { type: 'string', alias: 's', description: 'Path to scenario test directory (containing scenarios.json).  Cannot be combined with \'-f\'.' },
-			'remote-user': { type: 'string', alias: 'u', default: 'root', describe: 'Remote user', },
-			'collection-folder': { type: 'string', alias: 'c', default: '.', description: 'Path to folder containing \'src\' and \'test\' sub-folders.' },
-			'log-level': { choices: ['info' as 'info', 'debug' as 'debug', 'trace' as 'trace'], default: 'info' as 'info', description: 'Log level.' },
-			'quiet': { type: 'boolean', alias: 'q', default: false, description: 'Quiets output' },
-		})
-		.check(argv => {
-			if (argv['scenarios'] && argv['features']) {
-				throw new Error('Cannot combine --scenarios and --features');
-			}
-			return true;
-		});
-}
-
-export type FeaturesTestArgs = UnpackArgv<ReturnType<typeof featuresTestOptions>>;
-export interface FeaturesTestCommandInput {
-	cliHost: CLIHost;
-	pkg: PackageConfiguration;
-	baseImage: string;
-	collectionFolder: string;
-	features?: string[];
-	scenariosFolder: string | undefined;
-	remoteUser: string;
-	quiet: boolean;
-	logLevel: LogLevel;
-	disposables: (() => Promise<unknown> | undefined)[];
-}
-
-function featuresTestHandler(args: FeaturesTestArgs) {
-	(async () => await featuresTest(args))().catch(console.error);
-}
-
-async function featuresTest({
-	'base-image': baseImage,
-	'collection-folder': collectionFolder,
-	features,
-	scenarios: scenariosFolder,
-	'remote-user': remoteUser,
-	quiet,
-	'log-level': inputLogLevel,
-}: FeaturesTestArgs) {
-	const disposables: (() => Promise<unknown> | undefined)[] = [];
-	const dispose = async () => {
-		await Promise.all(disposables.map(d => d()));
-	};
-
-	const cwd = process.cwd();
-	const cliHost = await getCLIHost(cwd, loadNativeModule);
-	const extensionPath = path.join(__dirname, '..', '..');
-	const pkg = await getPackageConfig(extensionPath);
-
-	const logLevel = mapLogLevel(inputLogLevel);
-
-	const args: FeaturesTestCommandInput = {
-		baseImage,
-		cliHost,
-		logLevel,
-		quiet,
-		pkg,
-		collectionFolder: cliHost.path.resolve(collectionFolder),
-		features: features ? (Array.isArray(features) ? features as string[] : [features]) : undefined,
-		scenariosFolder: scenariosFolder ? cliHost.path.resolve(scenariosFolder) : undefined,
-		remoteUser,
-		disposables
-	};
-
-	const exitCode = await doFeaturesTestCommand(args);
-
-	await dispose();
-	process.exit(exitCode);
-}
-// -- End: 'features test' command
 
 function keyValuesToRecord(keyValues: string[]): Record<string, string> {
 	return keyValues.reduce((envs, env) => {
